requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.30
Pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
urllib3==2.2.1
